+++ CHANGELOG.md
@@ -15,10 +15,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Renamed concepts:
-  - "data package entities" -> "datasets"
-  - "data package instances" -> "data packages"
+  - "data package entity" -> "dataset"
+  - "data package instance" -> "data package"
+  - data package "descriptor" -> "spec"
   - datapackage.json -> dataset.json
-- `init`: Make source name a named argument (`-s`/`--source`). Change name option from `--package-name` to `--dataset`.
+- Update all CLI arguments to reflect the above. As a rule, `--spec` is used for spec inputs, `--dataset` for dataset reference inputs.
+- `init`: Make source name a named argument (`-s`/`--source`). Change name option from `--package-name` to `-n`/`--name`.
 - `build-package`: Update static code in all targets to redefine the
 `inPast` boolean operator in terms of computed relative time bounds.
 - `build-package`: Update `dpm_agent.proto` to include support for table joins. Update generated protobuf code in all targets.